+++ src/steps/driver_merge_pull_request_step.go
@@ -16,7 +16,7 @@ type DriverMergePullRequestStep struct {
 	enteredEmptyCommitMessage bool
 	mergeError                error
 	mergeSha                  string
-	PullRequestNumber         int64
+	PullRequestNumber         int
 }
 
 func (step *DriverMergePullRequestStep) CreateAbortStep() Step {
